Move geometry source control to BlockSource.tsx
